Moves token functions to token
