+++ clients/acl/client.go
@@ -1,10 +1,6 @@
 package acl
 
 import (
-	"bytes"
-	"encoding/json"
-	"errors"
-	"io/ioutil"
 	"net/http"
 	"os"
 	"strconv"
@@ -18,7 +14,7 @@ var log logger.Logging
 
 func init() {
 	// Add generic logger
-	log = logger.NewLogger("clients/oauth")
+	log = logger.NewLogger("clients/acl")
 }
 
 type AclClient interface {
@@ -48,50 +44,6 @@ func NewClient() *Client {
 	}
 
 	return &Client{BaseUrl: url, IsEnabled: isEnabled}
-}
-
-func (client Client) ExchangeToken(payload *ExchangePayload) (ExchangeResponse, error) {
-	// Transform Token struct to json payload
-	jsonValue, _ := json.Marshal(payload)
-	request, _ := http.NewRequest(http.MethodPost, client.BaseUrl+TOKEN_EXCHANGE_URL, bytes.NewBuffer(jsonValue))
-	log.Printf("Exchanging token for organisation %s", payload.Organisation)
-
-	// Set Headers for this request
-	request.Header.Set("Authorization", "Bearer "+payload.Token)
-	request.Header.Add("Content-Type", "application/json")
-
-	// Perform Request
-	res, wasRequested, err := client.HandleRequest(request)
-
-	// Check for errors, default evaluation is false
-	if err != nil {
-		log.Printf("Error processing Token exchange: %v", err)
-		return ExchangeResponse{}, err
-	}
-
-	// Check if the request was actually made
-	if !wasRequested {
-		// No request was made, let's return fake response
-		// This will be exactly the same token as requested for exchange
-		return ExchangeResponse{payload.Token}, nil
-	}
-
-	// read all response body
-	data, _ := ioutil.ReadAll(res.Body)
-
-	// print `data` as a string
-	log.Printf("%s\n", data)
-
-	// Success, populate token
-	if res.StatusCode == http.StatusOK {
-		result := ExchangeResponse{}
-		json.Unmarshal(data, &result)
-
-		// Return token
-		return result, nil
-	}
-
-	return ExchangeResponse{}, errors.New("error exchanging Token for new organisation")
 }
 
 // HandleRequest takes instance of the http.Request and performs request if client is enabled
